moved pdf logic from controller to service
